requirements.txt
# Core
streamlit>=1.32.0
numpy>=2.0.0
matplotlib>=3.7.0

# PDF text extraction
PyMuPDF>=1.23.0

# Local LLM
llama-cpp-python>=0.3.9

# Utility
diskcache>=5.6.3
jinja2>=3.1.2
typing-extensions>=4.5.0

# For form styling and animations
Pillow>=10.0.0
